+++ tasks/update-granules-cmr-metadata-file-links/index.js
@@ -3,6 +3,7 @@
 const cumulusMessageAdapter = require('@cumulus/cumulus-message-adapter-js');
 const get = require('lodash/get');
 const keyBy = require('lodash/keyBy');
+const isBoolean = require('lodash/isBoolean');
 const Logger = require('@cumulus/logger');
 const { getObjectSize } = require('@cumulus/aws-client/S3');
 const { s3 } = require('@cumulus/aws-client/services');
@@ -36,6 +37,8 @@ const logger = new Logger({ sender: '@cumulus/update-granules-cmr-metadata-file-
  * @param {Object} distributionBucketMap   - mapping of bucket->distribution path values
  *                                           (e.g. { bucket: distribution path })
  * @param {Object} excludeFileRegexPattern - pattern by which to exclude files from processing
+* @param {boolean} updateGranuleIdentifiers - Whether to update the Granule's Identifiers
+ *                                             and granuleUR
  * @returns {Promise<Object[]>} Array of updated CMR files with etags of newly updated files.
  *
  */
@@ -47,7 +50,8 @@ async function updateEachCmrFileMetadata(
   distEndpoint,
   bucketTypes,
   distributionBucketMap,
-  excludeFileRegexPattern
+  excludeFileRegexPattern,
+  updateGranuleIdentifiers
 ) {
   return await Promise.all(cmrFiles.map(async (cmrFile) => {
     const granuleId = cmrFile.granuleId;
@@ -71,7 +75,7 @@ async function updateEachCmrFileMetadata(
       bucketTypes,
       cmrGranuleUrlType,
       distributionBucketMap,
-      updateGranuleIdentifiers: true,
+      updateGranuleIdentifiers,
     });
   }));
 }
@@ -111,6 +115,8 @@ async function updateGranulesCmrMetadata(event) {
   const granules = event.input.granules.map((g) => addEtagsToFileObjects(g, incomingETags));
   const cmrFiles = granulesToCmrFileObjects(granules);
   const granulesByGranuleId = keyBy(granules, 'granuleId');
+  const updateGranuleIdentifiers = isBoolean(event.config.updateGranuleIdentifiers) ?
+    event.config.updateGranuleIdentifiers : true;
 
   const distributionBucketMap = await fetchDistributionBucketMap();
   const updatedCmrFiles = await updateEachCmrFileMetadata(
@@ -120,7 +126,8 @@ async function updateGranulesCmrMetadata(event) {
     config.distribution_endpoint,
     bucketTypes,
     distributionBucketMap,
-    config.excludeFileRegex
+    config.excludeFileRegex,
+    updateGranuleIdentifiers
   );
 
   const updatedGranulesByGranuleId = await updateCmrFileInfo(cmrFiles, granulesByGranuleId);